+++ configs/detection/dab_detr/dn_detr_r50_8x2_50e_coco.py
@@ -1,52 +1,7 @@
-_base_ = ['./dn_detr.py', './coco_detection.py', 'configs/base.py']
+_base_ = './dab_detr_r50_8x2_50e_coco.py'
 
-CLASSES = [
-    'person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train',
-    'truck', 'boat', 'traffic light', 'fire hydrant', 'stop sign',
-    'parking meter', 'bench', 'bird', 'cat', 'dog', 'horse', 'sheep', 'cow',
-    'elephant', 'bear', 'zebra', 'giraffe', 'backpack', 'umbrella', 'handbag',
-    'tie', 'suitcase', 'frisbee', 'skis', 'snowboard', 'sports ball', 'kite',
-    'baseball bat', 'baseball glove', 'skateboard', 'surfboard',
-    'tennis racket', 'bottle', 'wine glass', 'cup', 'fork', 'knife', 'spoon',
-    'bowl', 'banana', 'apple', 'sandwich', 'orange', 'broccoli', 'carrot',
-    'hot dog', 'pizza', 'donut', 'cake', 'chair', 'couch', 'potted plant',
-    'bed', 'dining table', 'toilet', 'tv', 'laptop', 'mouse', 'remote',
-    'keyboard', 'cell phone', 'microwave', 'oven', 'toaster', 'sink',
-    'refrigerator', 'book', 'clock', 'vase', 'scissors', 'teddy bear',
-    'hair drier', 'toothbrush'
-]
-
-log_config = dict(
-    interval=50,
-    hooks=[
-        dict(type='TextLoggerHook'),
-        # dict(type='TensorboardLoggerHook')
-    ])
-
-checkpoint_config = dict(interval=10)
-# optimizer
-paramwise_options = {'backbone': dict(lr_mult=0.1, weight_decay_mult=1.0)}
-optimizer = dict(
-    type='AdamW',
-    lr=1e-4,
-    weight_decay=1e-4,
-    paramwise_options=paramwise_options)
-optimizer_config = dict(grad_clip=dict(max_norm=0.1, norm_type=2))
-# learning policy
-lr_config = dict(policy='step', step=[40])
-
-total_epochs = 50
-
-# evaluation
-eval_config = dict(initial=True, interval=1, gpu_collect=False)
-# eval_config = dict(interval=1, gpu_collect=False)
-eval_pipelines = [
-    dict(
-        mode='test',
-        evaluators=[
-            dict(type='CocoDetectionEvaluator', classes=CLASSES),
-        ],
-    )
-]
-
-find_unused_parameters = False
+# model settings
+model = dict(
+    head=dict(
+        dn_components=dict(
+            scalar=5, label_noise_scale=0.2, box_noise_scale=0.4)))
