+++ easycv/models/detection/dab_detr/dab_detr_head.py
@@ -8,14 +8,10 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from easycv.models.builder import HEADS, build_neck
-from easycv.models.detection.utils import (MLP, HungarianMatcher, accuracy,
+from easycv.models.detection.utils import (MLP, HungarianMatcher, SetCriterion,
                                            box_cxcywh_to_xyxy,
-                                           box_xyxy_to_cxcywh,
-                                           generalized_box_iou,
-                                           inverse_sigmoid)
-from easycv.models.loss.focal_loss import py_sigmoid_focal_loss
-from easycv.models.utils import get_world_size, is_dist_avail_and_initialized
-from .dn_components import compute_dn_loss, dn_post_process, prepare_for_dn
+                                           box_xyxy_to_cxcywh, inverse_sigmoid)
+from .dn_components import dn_post_process, prepare_for_dn
 
 
 @HEADS.register_module()
@@ -26,8 +22,6 @@ class DABDETRHead(nn.Module):
     Args:
         num_classes (int): Number of categories excluding the background.
     """
-
-    _version = 2
 
     def __init__(self,
                  num_classes,
@@ -61,7 +55,8 @@ class DABDETRHead(nn.Module):
             num_classes,
             matcher=self.matcher,
             weight_dict=weight_dict,
-            losses=['labels', 'boxes', 'cardinality'])
+            losses=['labels', 'boxes', 'cardinality'],
+            loss_class_type='focal_loss')
         self.postprocess = PostProcess(num_select=num_select)
         self.transformer = build_neck(transformer)
 
@@ -137,10 +132,12 @@ class DABDETRHead(nn.Module):
             return query_embed, tgt, attn_mask, mask_dict
         else:
             query_embed = query_embed.unsqueeze(1).repeat(1, bs, 1)
-            num_queries = query_embed.shape[0]
             if self.num_patterns == 0:
                 tgt = torch.zeros(
-                    num_queries, bs, self.d_model, device=query_embed.device)
+                    self.num_queries,
+                    bs,
+                    self.embed_dims,
+                    device=query_embed.device)
             else:
                 tgt = self.patterns.weight[:, None, None, :].repeat(
                     1, self.num_queries, bs,
@@ -329,202 +326,3 @@ class PostProcess(nn.Module):
         }
 
         return results
-
-
-class SetCriterion(nn.Module):
-    """ This class computes the loss for Conditional DETR.
-    The process happens in two steps:
-        1) we compute hungarian assignment between ground truth boxes and the outputs of the model
-        2) we supervise each pair of matched ground-truth / prediction (supervise class and box)
-    """
-
-    def __init__(self, num_classes, matcher, weight_dict, losses):
-        """ Create the criterion.
-        Parameters:
-            num_classes: number of object categories, omitting the special no-object category
-            matcher: module able to compute a matching between targets and proposals
-            weight_dict: dict containing as key the names of the losses and as values their relative weight.
-            losses: list of all the losses to be applied. See get_loss for list of available losses.
-        """
-        super().__init__()
-        self.num_classes = num_classes
-        self.matcher = matcher
-        self.weight_dict = weight_dict
-        self.losses = losses
-
-    def loss_labels(self, outputs, targets, indices, num_boxes, log=True):
-        """Classification loss (Binary focal loss)
-        targets dicts must contain the key "labels" containing a tensor of dim [nb_target_boxes]
-        """
-        assert 'pred_logits' in outputs
-        src_logits = outputs['pred_logits']
-
-        idx = self._get_src_permutation_idx(indices)
-        target_classes_o = torch.cat(
-            [t['labels'][J] for t, (_, J) in zip(targets, indices)])
-        target_classes = torch.full(
-            src_logits.shape[:2],
-            self.num_classes,
-            dtype=torch.int64,
-            device=src_logits.device)
-        target_classes[idx] = target_classes_o
-
-        target_classes_onehot = torch.zeros([
-            src_logits.shape[0], src_logits.shape[1], src_logits.shape[2] + 1
-        ],
-                                            dtype=src_logits.dtype,
-                                            layout=src_logits.layout,
-                                            device=src_logits.device)
-        target_classes_onehot.scatter_(2, target_classes.unsqueeze(-1), 1)
-
-        target_classes_onehot = target_classes_onehot[:, :, :-1]
-        loss_ce = py_sigmoid_focal_loss(
-            src_logits,
-            target_classes_onehot.long(),
-            alpha=0.25,
-            gamma=2,
-            reduction='none').mean(1).sum() / num_boxes
-        loss_ce = loss_ce * src_logits.shape[1] * self.weight_dict['loss_ce']
-        losses = {'loss_ce': loss_ce}
-
-        if log:
-            # TODO this should probably be a separate loss, not hacked in this one here
-            losses['class_error'] = 100 - accuracy(src_logits[idx],
-                                                   target_classes_o)[0]
-        return losses
-
-    @torch.no_grad()
-    def loss_cardinality(self, outputs, targets, indices, num_boxes):
-        """ Compute the cardinality error, ie the absolute error in the number of predicted non-empty boxes
-        This is not really a loss, it is intended for logging purposes only. It doesn't propagate gradients
-        """
-        pred_logits = outputs['pred_logits']
-        device = pred_logits.device
-        tgt_lengths = torch.as_tensor([len(v['labels']) for v in targets],
-                                      device=device)
-        # Count the number of predictions that are NOT "no-object" (which is the last class)
-        card_pred = (pred_logits.argmax(-1) !=
-                     pred_logits.shape[-1] - 1).sum(1)
-        card_err = F.l1_loss(card_pred.float(), tgt_lengths.float())
-        losses = {'cardinality_error': card_err}
-        return losses
-
-    def loss_boxes(self, outputs, targets, indices, num_boxes):
-        """Compute the losses related to the bounding boxes, the L1 regression loss and the GIoU loss
-           targets dicts must contain the key "boxes" containing a tensor of dim [nb_target_boxes, 4]
-           The target boxes are expected in format (center_x, center_y, w, h), normalized by the image size.
-        """
-        assert 'pred_boxes' in outputs
-        idx = self._get_src_permutation_idx(indices)
-        src_boxes = outputs['pred_boxes'][idx]
-        target_boxes = torch.cat(
-            [t['boxes'][i] for t, (_, i) in zip(targets, indices)], dim=0)
-
-        loss_bbox = F.l1_loss(src_boxes, target_boxes, reduction='none')
-
-        losses = {}
-        losses['loss_bbox'] = loss_bbox.sum(
-        ) / num_boxes * self.weight_dict['loss_bbox']
-
-        loss_giou = 1 - torch.diag(
-            generalized_box_iou(
-                box_cxcywh_to_xyxy(src_boxes),
-                box_cxcywh_to_xyxy(target_boxes)))
-        losses['loss_giou'] = loss_giou.sum(
-        ) / num_boxes * self.weight_dict['loss_giou']
-
-        return losses
-
-    def _get_src_permutation_idx(self, indices):
-        # permute predictions following indices
-        batch_idx = torch.cat(
-            [torch.full_like(src, i) for i, (src, _) in enumerate(indices)])
-        src_idx = torch.cat([src for (src, _) in indices])
-        return batch_idx, src_idx
-
-    def _get_tgt_permutation_idx(self, indices):
-        # permute targets following indices
-        batch_idx = torch.cat(
-            [torch.full_like(tgt, i) for i, (_, tgt) in enumerate(indices)])
-        tgt_idx = torch.cat([tgt for (_, tgt) in indices])
-        return batch_idx, tgt_idx
-
-    def get_loss(self, loss, outputs, targets, indices, num_boxes, **kwargs):
-        loss_map = {
-            'labels': self.loss_labels,
-            'cardinality': self.loss_cardinality,
-            'boxes': self.loss_boxes,
-        }
-        assert loss in loss_map, f'do you really want to compute {loss} loss?'
-        return loss_map[loss](outputs, targets, indices, num_boxes, **kwargs)
-
-    def forward(self, outputs, targets, mask_dict=None, return_indices=False):
-        """ This performs the loss computation.
-        Parameters:
-             outputs: dict of tensors, see the output specification of the model for the format
-             targets: list of dicts, such that len(targets) == batch_size.
-                      The expected keys in each dict depends on the losses applied, see each loss' doc
-
-             return_indices: used for vis. if True, the layer0-5 indices will be returned as well.
-        """
-
-        outputs_without_aux = {
-            k: v
-            for k, v in outputs.items() if k != 'aux_outputs'
-        }
-
-        # Retrieve the matching between the outputs of the last layer and the targets
-        indices = self.matcher(outputs_without_aux, targets)
-        if return_indices:
-            indices0_copy = indices
-            indices_list = []
-
-        # Compute the average number of target boxes accross all nodes, for normalization purposes
-        num_boxes = sum(len(t['labels']) for t in targets)
-        num_boxes = torch.as_tensor([num_boxes],
-                                    dtype=torch.float,
-                                    device=next(iter(outputs.values())).device)
-        if is_dist_avail_and_initialized():
-            torch.distributed.all_reduce(num_boxes)
-        num_boxes = torch.clamp(num_boxes / get_world_size(), min=1).item()
-
-        # Compute all the requested losses
-        losses = {}
-        for loss in self.losses:
-            losses.update(
-                self.get_loss(loss, outputs, targets, indices, num_boxes))
-
-        # In case of auxiliary losses, we repeat this process with the output of each intermediate layer.
-        if 'aux_outputs' in outputs:
-            for i, aux_outputs in enumerate(outputs['aux_outputs']):
-                indices = self.matcher(aux_outputs, targets)
-                if return_indices:
-                    indices_list.append(indices)
-                for loss in self.losses:
-                    if loss == 'masks':
-                        # Intermediate masks losses are too costly to compute, we ignore them.
-                        continue
-                    kwargs = {}
-                    if loss == 'labels':
-                        # Logging is enabled only for the last layer
-                        kwargs = {'log': False}
-                    l_dict = self.get_loss(loss, aux_outputs, targets, indices,
-                                           num_boxes, **kwargs)
-                    l_dict = {k + f'_{i}': v for k, v in l_dict.items()}
-                    losses.update(l_dict)
-
-        if mask_dict is not None:
-            # print('dn_loss!!!!!!!!!!')
-            # dn loss computation
-            aux_num = 0
-            if 'aux_outputs' in outputs:
-                aux_num = len(outputs['aux_outputs'])
-            dn_losses = compute_dn_loss(mask_dict, self.training, aux_num,
-                                        0.25)
-            losses.update(dn_losses)
-
-        if return_indices:
-            indices_list.append(indices0_copy)
-            return losses, indices_list
-
-        return losses
